bots/ken/requirements.txt
re-ircbot==1.7.0
cleverbot-scraper==0.2.7
